Update Eth-DM specs as per PoC
This is a first version that has a number of limitations.
For example, Bob does not know the origin of the messages and Bob's
Ethereum address is attached to all his messages.
We intend to resolve this limitation in future versions.

Note that the encryption is done at application level, we do not use the
`version` field of `WakuMessage`.